pyproject.toml
[project]
name = "dabs-with-uv"
dynamic = ["version"]
description = "Example project with Databicks Asset Bundles and uv"
readme = "README.md"
authors = [
    { name = "renardeinside", email = "polarpersonal@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.10.4",
]

[project.scripts]
aggregator = "dabs_with_uv.aggregator:entrypoint"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "databricks-connect<16",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.4",
]

[tool.hatch.version]
source = "vcs"
